Accept mode param in image size definition
This modification adds the possibility to add a new parameter in the images size definition: mode.

If mode = 'cropCenter', the image will be cropped to match the specified dimensions, no matter the original aspect ratio.
If not specified the default behavior will be applied.

This is handy if you need to have different aspect ratios in the same image, by cropping them centered. If the image is smaller than the required dimensions, it will be amplified.